Move cmdoptions to pip._internal.cli
