+++ src/pip/_internal/__init__.py
@@ -39,7 +39,7 @@ else:
             securetransport.inject_into_urllib3()
 
 from pip import __version__
-from pip._internal import cmdoptions
+from pip._internal.cli import cmdoptions
 from pip._internal.exceptions import CommandError, PipError
 from pip._internal.utils.misc import get_installed_distributions, get_prog
 from pip._internal.utils import deprecation
