+++ src/pip/_internal/commands/wheel.py
@@ -4,7 +4,7 @@ from __future__ import absolute_import
 import logging
 import os
 
-from pip._internal import cmdoptions
+from pip._internal.cli import cmdoptions
 from pip._internal.cli.basecommand import RequirementCommand
 from pip._internal.cache import WheelCache
 from pip._internal.exceptions import CommandError, PreviousBuildDirError
